+++ packages/stream_chat_flutter/lib/src/attachment/attachment.dart
@@ -1,0 +1,6 @@
+export 'file_attachment.dart';
+export 'giphy_attachment.dart';
+export 'image_attachment.dart';
+export 'video_attachment.dart';
+export 'attachment_widget.dart'
+    show AttachmentError, AttachmentSource, AttachmentSourceX;