+++ packages/stream_chat_flutter/lib/src/attachment/attachment_title.dart
@@ -1,8 +1,8 @@
 import 'package:flutter/material.dart';
 import 'package:stream_chat_flutter_core/stream_chat_flutter_core.dart';
 
-import 'stream_chat_theme.dart';
-import 'utils.dart';
+import '../stream_chat_theme.dart';
+import '../utils.dart';
 
 class AttachmentTitle extends StatelessWidget {
   const AttachmentTitle({