+++ packages/stream_chat_flutter/lib/src/message_reactions_modal.dart
@@ -23,7 +23,6 @@ class MessageReactionsModal extends StatelessWidget {
   final ShapeBorder messageShape;
   final ShapeBorder attachmentShape;
   final void Function(User) onUserAvatarTap;
-  final Map<String, VideoPackage> videoPackages;
 
   const MessageReactionsModal({
     Key key,
@@ -37,7 +36,6 @@ class MessageReactionsModal extends StatelessWidget {
     this.reverse = false,
     this.showUserAvatar = DisplayWidget.show,
     this.onUserAvatarTap,
-    this.videoPackages,
   }) : super(key: key);
 
   @override
@@ -146,7 +144,6 @@ class MessageReactionsModal extends StatelessWidget {
                                   (message.status ==
                                           MessageSendingStatus.sent ||
                                       message.status == null),
-                              videoPackages: videoPackages,
                             ),
                           ),
                           if (message.latestReactions?.isNotEmpty == true) ...[